moved logic from the main card-search-result component
